ProjectM-BankStatement-Extractor/requirements.txt
ocr-nanonets-wrapper
pandas